+++ commands/index.js
@@ -1,6 +1,7 @@
 const {accountingCategories} = require('./constants');
 const callRawApi = require('./call_raw_api');
 const {exchanges} = require('./constants');
+const fetchRequest = require('./fetch_request');
 const {marketPairs} = require('./constants');
 const {peerSortOptions} = require('./constants');
 const {rateProviders} = require('./constants');
@@ -10,6 +11,7 @@ module.exports = {
   accountingCategories,
   callRawApi,
   exchanges,
+  fetchRequest,
   marketPairs,
   peerSortOptions,
   rateProviders,